re-write message handling
Now, two messages are received:
The first message is a 4 byte header containing the length of the actual
payload.
The buffer size for the second message is then set to this specific
size.
Then, the second message containing the actual payload is received and
passed to the parser.
No more fiddling around with ring buffers and stop characters.